Replaced docks with tab widgets and correct desktop ratio calculation

+++ range/Range/src/gl_widget.h
@@ -33,7 +33,7 @@ class GLWidget : public QOpenGLWidget
         //! Local axis viewport height.
         static const GLsizei lAxisWpHeight;
         //! Desktop device pixel ratio.
-        int desktopDevicePixelRatio;
+        qreal desktopDevicePixelRatio;
 
     protected:
